+++ Player.h
@@ -12,13 +12,13 @@
 class Player final : public Entity, public CommandListener {
     Texture2D m_walkSprites{};          // Walk sprite sheet
     Rectangle m_spriteRect{};           // Rectangle defining the current sprite
-    const int m_numFrames;              // Number of sprites in sheet width-wise
-    int m_frameIndex;                   // Currently selected sprite
-    const int m_frameDelayAmount;       // Duration to delay next frame
-    int m_frameDelayClock;              // Clock counting up to amount
+    const uint8_t m_numFrames;          // Number of sprites in sheet width-wise
+    uint8_t m_frameIndex;               // Currently selected sprite
+    const uint8_t m_frameDelayAmount;   // Duration to delay next frame
+    uint8_t m_frameDelayClock;              // Clock counting up to amount
     bool m_moving;                      // Whether or not the player is moving
     bool m_onGround;                    // Whether the player is on the ground or not
-    int m_lastDirection;                // Last direction the player was moving for animation purposes
+    uint8_t m_lastDirection;                // Last direction the player was moving for animation purposes
     b2Polygon m_footpawSensorBox{};     // Footpaw sensor box polygon
     b2ShapeDef m_footpawSensorShape{};  // Footpaw sensor box shape
     b2ShapeId m_footpawSensorId{};      // Footpaw sensor box shape ID
